ahora funciona ir por el centro

+++ src/warrior_01/include/warrior_01/warrior.hpp
@@ -27,7 +27,7 @@ public:
     void perform_movement(bool MOVE_TO_GOAL, bool FOLLOW_WALL, float close_enough_avoiding, float below_avoiding, float fixing_direction_avoiding, float front_distance);
 
 
-    int encontrarCercanoNoObstaculo(const std::vector<int>& obstacles, int search_index, const sensor_msgs::msg::LaserScan::SharedPtr msg);
+    float encontrarCercanoNoObstaculo(const std::vector<int>& obstacles, int search_index, const sensor_msgs::msg::LaserScan::SharedPtr msg);
 
     float compute_euclidean_distance(float x1, float x2, float y1, float y2);
